chore(server): refactor HTTP transport factory to avoid circular dependency
